backend/main.py
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
import uvicorn
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(title="EcoArbitrage AI - Agentic Guard")

# Enable CORS for frontend communication
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class AnalysisRequest(BaseModel):
    region: str
    sector: str
    data_points: list

@app.get("/")
async def root():
    return {
        "message": "EcoArbitrage AI API is active",
        "developer": "Raphasha27",
        "status": "Operational",
        "version": "1.0.0"
    }

@app.get("/api/dashboard")
async def get_dashboard_data():
    return {
        "metrics": {
            "economic_stability": 78,
            "social_impact_score": 89,
            "ai_intervention_count": 254,
            "arbitrage_efficiency": "94.2%"
        },
        "incidents": [
            {"id": 1, "title": "Load Shedding Mitigation AI", "status": "Optimizing", "severity": "High", "location": "Gauteng Grid"},
            {"id": 2, "title": "Agricultural Yield Arbitrage", "status": "Active", "severity": "Medium", "location": "Free State"},
            {"id": 3, "title": "Taxi Industry Route Optimization", "status": "Stable", "severity": "Low", "location": "Western Cape"},
            {"id": 4, "title": "Mining Safety Sentinel active", "status": "Monitoring", "severity": "Medium", "location": "North West"}
        ]
    }

@app.get("/api/global-feed")
async def get_global_feed():
    return [
        {"id": 101, "event": "Fair Fuel Pricing deployed in Durban Harbor", "time": "2 mins ago"},
        {"id": 102, "event": "Resource Maldistribution corrected in Polokwane Clinics", "time": "15 mins ago"},
        {"id": 103, "event": "Solar Micro-grid optimized for Khayelitsha community", "time": "1 hr ago"},
        {"id": 104, "event": "Educational subsidy AI active in Eastern Cape schools", "time": "3 hrs ago"}
    ]

@app.post("/api/initiate-guard")
async def initiate_guard():
    return {
        "status": "Success",
        "action": "Mzansi Guard Deployed",
        "timestamp": "2025-12-31T12:00:00Z",
        "protection_layer": "Ubuntu Shield v2.1",
        "message": "Autonomous Market Stabilization Active across South Africa."
    }

@app.get("/api/ai-reasoning")
async def get_ai_reasoning():
    # Purpose-Driven Solutions with Global Integration
    solutions = [
        {
            "problem": "Education Disparity (Eastern Cape)",
            "solution": "Deploying P2P Agentic Learning Meshes and Global Knowledge Bridging.",
            "purpose": "Averting generational poverty through decentralized education.",
            "impact_prediction": "+45% literacy rate in 12 months",
            "global_status": "Syncing with UNESCO Standards"
        },
        {
            "problem": "Healthcare Inequality (Limpopo)",
            "solution": "Autonomous Mobile Clinic Routing & Global Bio-Alert Integration.",
            "purpose": "Universal health access via AI-managed logistics.",
            "impact_prediction": "98% coverage of rural sectors",
            "global_status": "Linked to WHO Early Warning System"
        }
    ]
    return {
        "engine": "Socio-Reasoner v2.4",
        "current_focus": "Socio-Economic Fairness Pillar",
        "active_solutions": solutions
    }

@app.post("/api/analyze")
async def analyze_impact(request: AnalysisRequest):
    # This would integrate with an AI model (OpenAI/Gemini)
    # Placeholder for AI logic
    return {
        "region": request.region,
        "sector": request.sector,
        "ai_recommendation": f"Optimization for {request.sector} in {request.region} initiated. Agentic Guard active.",
        "confidence_score": 0.98
    }

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)
